requirements.txt
Flask
pg8000
python-dotenv
Werkzeug
Pillow
pandas
openpyxl
